requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
rdflib==7.2.1
pyshacl==0.30.1
click==8.3.0
PyYAML==6.0.2
pytest==8.3.3